refactor(occasion): refactor data models and repository structure
- Removed the old OccasionsEntity and OccasionEntity classes, replacing them with OccasionResponseEntity and OccasionEntity for better clarity.
- Updated the DTOs for occasions and products to include new metadata handling.
- Refactored the OccasionRemoteDataSource to use a contract interface for better abstraction.
- Implemented OccasionRemoteDataSourceImpl to handle API calls for occasions and products.
- Adjusted the use cases and repository implementations to align with the new data structure.
- Enhanced the OccasionCubit to manage state more effectively with the new entity structure.
- Updated the UI components to reflect changes in the data models and ensure proper state management.
