+++ lib/features/occasions/data/data_source/contract/occasion_remote_data_source.dart
@@ -1,0 +1,8 @@
+import '../../model/occasions_response_dto.dart';
+import '../../model/products_response_dto.dart';
+
+abstract class OccasionRemoteDataSource {
+  Future<OccasionsResponseDto> getAllOccasions();
+
+  Future<ProductsResponseDto> getProductsByOccasion(String occasionId);
+}
