+++ lib/features/occasions/data/model/products_response_dto.dart
@@ -1,6 +1,7 @@
 import 'package:json_annotation/json_annotation.dart';
 
-import '../../../../core/utils/shared_models/product_entity.dart';
+import 'meta_data_dto.dart';
+import 'product_dto.dart';
 
 part 'products_response_dto.g.dart';
 
@@ -9,9 +10,9 @@ class ProductsResponseDto {
   @JsonKey(name: "message")
   final String? message;
   @JsonKey(name: "metadata")
-  final Metadata? metadata;
+  final MetadataDto? metadata;
   @JsonKey(name: "products")
-  final List<Products>? products;
+  final List<ProductDto>? products;
 
   ProductsResponseDto({this.message, this.metadata, this.products});
 
@@ -21,122 +22,5 @@ class ProductsResponseDto {
 
   Map<String, dynamic> toJson() {
     return _$ProductsResponseDtoToJson(this);
-  }
-}
-
-@JsonSerializable()
-class Metadata {
-  @JsonKey(name: "currentPage")
-  final int? currentPage;
-  @JsonKey(name: "totalPages")
-  final int? totalPages;
-  @JsonKey(name: "limit")
-  final int? limit;
-  @JsonKey(name: "totalItems")
-  final int? totalItems;
-
-  Metadata({this.currentPage, this.totalPages, this.limit, this.totalItems});
-
-  factory Metadata.fromJson(Map<String, dynamic> json) {
-    return _$MetadataFromJson(json);
-  }
-
-  Map<String, dynamic> toJson() {
-    return _$MetadataToJson(this);
-  }
-}
-
-@JsonSerializable()
-class Products {
-  @JsonKey(name: "_id")
-  final String? Id;
-  @JsonKey(name: "title")
-  final String? title;
-  @JsonKey(name: "slug")
-  final String? slug;
-  @JsonKey(name: "description")
-  final String? description;
-  @JsonKey(name: "imgCover")
-  final String? imgCover;
-  @JsonKey(name: "images")
-  final List<String>? images;
-  @JsonKey(name: "price")
-  final int? price;
-  @JsonKey(name: "priceAfterDiscount")
-  final int? priceAfterDiscount;
-  @JsonKey(name: "quantity")
-  final int? quantity;
-  @JsonKey(name: "category")
-  final String? category;
-  @JsonKey(name: "occasion")
-  final String? occasion;
-  @JsonKey(name: "createdAt")
-  final String? createdAt;
-  @JsonKey(name: "updatedAt")
-  final String? updatedAt;
-  @JsonKey(name: "__v")
-  final int? V;
-  @JsonKey(name: "discount")
-  final int? discount;
-  @JsonKey(name: "sold")
-  final int? sold;
-  @JsonKey(name: "rateAvg")
-  final double? rateAvg;
-  @JsonKey(name: "rateCount")
-  final int? rateCount;
-  @JsonKey(name: "id")
-  final String? id;
-
-  Products({
-    this.Id,
-    this.title,
-    this.slug,
-    this.description,
-    this.imgCover,
-    this.images,
-    this.price,
-    this.priceAfterDiscount,
-    this.quantity,
-    this.category,
-    this.occasion,
-    this.createdAt,
-    this.updatedAt,
-    this.V,
-    this.discount,
-    this.sold,
-    this.rateAvg,
-    this.rateCount,
-    this.id,
-  });
-
-  factory Products.fromJson(Map<String, dynamic> json) {
-    return _$ProductsFromJson(json);
-  }
-
-  Map<String, dynamic> toJson() {
-    return _$ProductsToJson(this);
-  }
-
-  ProductEntity toEntity() {
-    return ProductEntity(
-      title: title,
-      description: description,
-      imgCover: imgCover,
-      images: images,
-      price: price,
-      priceAfterDiscount: priceAfterDiscount,
-      quantity: quantity,
-      discount: discount,
-      sold: sold,
-      rateAvg: rateAvg,
-      rateCount: rateCount,
-      productId: id,
-      category: category,
-      occasion: occasion,
-      createdAt: DateTime.tryParse(createdAt ?? ''),
-      updatedAt: DateTime.tryParse(updatedAt ?? ''),
-      v: V,
-      slug: slug,
-    );
   }
 }
