+++ lib/features/occasions/domain/entity/occasion_entity.dart
@@ -1,0 +1,6 @@
+class OccasionEntity {
+  final String? id;
+  final String? name;
+
+  OccasionEntity({required this.id, required this.name});
+}
