+++ lib/features/occasions/domain/entity/occasions_response_entity.dart
@@ -1,0 +1,8 @@
+import 'occasion_entity.dart';
+
+class OccasionResponseEntity {
+  final String? message;
+  final List<OccasionEntity>? occasions;
+
+  OccasionResponseEntity({this.message, this.occasions});
+}
