+++ lib/features/occasions/domain/usecase/get_all_occasions_usecase.dart
@@ -1,5 +1,5 @@
 import 'package:ecommerce_flower_app/core/utils/datasource_excution/api_result.dart';
-import 'package:ecommerce_flower_app/features/occasions/domain/entity/occasions_entity.dart';
+import 'package:ecommerce_flower_app/features/occasions/domain/entity/occasions_response_entity.dart';
 import 'package:ecommerce_flower_app/features/occasions/domain/repo/occasion_repo.dart';
 import 'package:injectable/injectable.dart';
 
@@ -8,7 +8,7 @@ class GetAllOccasionsUsecase {
   final OccasionRepo _occasionRepo;
   GetAllOccasionsUsecase(this._occasionRepo);
 
-  Future<Result<OccasionsEntity>> call() async {
+  Future<Result<OccasionResponseEntity>> call() async {
     return await _occasionRepo.getAllOccasions();
   }
 }
