+++ lib/features/occasions/presentation/view/screen/occasion_screen.dart
@@ -2,24 +2,33 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:easy_localization/easy_localization.dart';
 
-import '../../../../../core/assets/app_colors.dart';
 import '../../../../../core/utils/di/di.dart';
 import '../../../../../core/utils/l10n/locale_keys.g.dart';
 import '../../view_model/occasion_cubit.dart';
 import '../widgets/custom_occasion_tab_bar.dart';
 import '../widgets/product_view.dart';
 
-class OccasionScreen extends StatelessWidget {
+class OccasionScreen extends StatefulWidget {
   final int? occasionIndex;
   const OccasionScreen({super.key, this.occasionIndex});
 
   @override
+  State<OccasionScreen> createState() => _OccasionScreenState();
+}
+
+class _OccasionScreenState extends State<OccasionScreen> {
+  final viewModel = getIt<OccasionCubit>();
+
+  @override
+  void initState() {
+    super.initState();
+    viewModel.doIntent(OccasionRequestAction(index: widget.occasionIndex));
+  }
+
+  @override
   Widget build(BuildContext context) {
-    final viewModel = getIt<OccasionCubit>();
     return BlocProvider(
-      create:
-          (_) =>
-              viewModel..doIntent(OccasionRequestAction(index: occasionIndex)),
+      create: (_) => viewModel,
       child: BlocBuilder<OccasionCubit, OccasionState>(
         builder: (context, state) {
           final cubit = context.read<OccasionCubit>();
