+++ lib/features/occasions/presentation/view_model/occasion_cubit.dart
@@ -1,15 +1,15 @@
-import 'package:bloc/bloc.dart';
-import 'package:ecommerce_flower_app/features/occasions/domain/usecase/get_products_by_id_usecase.dart';
 import 'package:equatable/equatable.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:injectable/injectable.dart';
-import 'package:logger/logger.dart';
 
 import '../../../../core/base/base_state.dart';
 import '../../../../core/utils/datasource_excution/api_result.dart';
 import '../../../../core/utils/shared_models/product_entity.dart';
-import '../../domain/entity/occasions_entity.dart';
+import '../../domain/entity/occasion_entity.dart';
+import '../../domain/entity/occasions_response_entity.dart';
 import '../../domain/usecase/get_all_occasions_usecase.dart';
+import '../../domain/usecase/get_products_by_id_usecase.dart';
 
 part 'occasion_state.dart';
 
@@ -30,20 +30,23 @@ class OccasionCubit extends Cubit<OccasionState> {
     emit(state.copyWith(baseState: BaseLoadingState()));
     final result = await _getAllOccasionsUsecase.call();
     switch (result) {
-      case SuccessResult<OccasionsEntity>():
-        occasions = result.data.occasions;
-        myTabs = occasions?.map((e) => Tab(text: e.name)).toList();
-        emit(state.copyWith(baseState: BaseSuccessState()));
-        _changeOccasionTab((index ?? 1) - 1);
-        break;
-      case FailureResult<OccasionsEntity>():
-        emit(
-          state.copyWith(
-            baseState: BaseErrorState(
-              errorMessage: result.exception.toString(),
+      case SuccessResult<OccasionResponseEntity>():
+        {
+          occasions = result.data.occasions;
+          myTabs = occasions?.map((e) => Tab(text: e.name)).toList();
+          emit(state.copyWith(baseState: BaseSuccessState()));
+          _changeOccasionTab((index ?? 1) - 1);
+        }
+      case FailureResult<OccasionResponseEntity>():
+        {
+          emit(
+            state.copyWith(
+              baseState: BaseErrorState(
+                errorMessage: result.exception.toString(),
+              ),
             ),
-          ),
-        );
+          );
+        }
     }
   }
 
@@ -52,17 +55,20 @@ class OccasionCubit extends Cubit<OccasionState> {
     final result = await _getProductsByIdUsecase.call(occasionId);
     switch (result) {
       case SuccessResult<List<ProductEntity>>():
-        emit(state.copyWith(baseState: BaseSuccessState()));
-        products = result.data;
-        Logger().f("Products: $products");
+        {
+          emit(state.copyWith(baseState: BaseSuccessState()));
+          products = result.data;
+        }
       case FailureResult<List<ProductEntity>>():
-        emit(
-          state.copyWith(
-            baseState: BaseErrorState(
-              errorMessage: result.exception.toString(),
+        {
+          emit(
+            state.copyWith(
+              baseState: BaseErrorState(
+                errorMessage: result.exception.toString(),
+              ),
             ),
-          ),
-        );
+          );
+        }
     }
   }
 
@@ -77,11 +83,17 @@ class OccasionCubit extends Cubit<OccasionState> {
   void doIntent(OccasionAction action) async {
     switch (action) {
       case OccasionRequestAction():
-        _getAllOccasions(action.index);
+        {
+          _getAllOccasions(action.index);
+        }
       case ProductsRequestAction():
-        _getProductsById(action.occasionId);
+        {
+          _getProductsById(action.occasionId);
+        }
       case ChangeOccasionTabAction():
-        _changeOccasionTab(action.tabIndex);
+        {
+          _changeOccasionTab(action.tabIndex);
+        }
     }
   }
 }
